refactor catalogue test
